SELinux: Solve "Could not set context for /init" problem
The init used to panic when executing restorecon, logging "Could not set
context for /init: Operation not supported on transport endpoint".

That's due to that by default, SELinux doesn't support setting contexts
for files in rootfs.

Luckily, the NSA's developers gave a patch to fix it. They improved the
hook module to add support for rootfs file context setting. After applying
the patch, the init can be successfully proceeded, and goes into the second
stage.

Related links:

- A programmer's same issue on NSA's bug tracker: http://seandroid-list.tycho.nsa.narkive.com/D3qnO4XY/restorecon-init-fail-in-the-init-cpp-file
- NSA's kernel patch: https://android-review.googlesource.com/c/kernel/common/+/58360/1