c8816_defconfig: Add c8816_defconfig_twrp
I added a new config called c8816_defconfig_twrp, which is dedicated to
TWRP.

Because C8816 only provides a 16MiB partition of TWRP, both DTS and
Kernel should be pruned in order to save spaces as possible.

As some drivers and components are definitely not used in TWRP, I pruned
them to build a smaller kernel, and save space.

Pruned components:
- NFC
- Bluetooth (BT)
- Camera
- Remote controller (RC)
- Video capture (Video4Linux)

For more details, simply compare the 2 config files: c8816_defconfig and
c8816_defconfig_twrp.